Switch per-app volume, rating and previous/next track to F11/F12 from F10/F11.
To be honest, I have not used these in years since I originally developed StreamVision with the IBM M15 keyboard. These days, I am using keyboards with layouts similar to the Apple Bluetooth layout, where volume up/down are on F11 and F12. Having related functions offset is just confusing.

The reason I have not moved track switching to F7 and F9 (where they are on my MBP keyboard) are that the play/pause button is in different places on the two keyboards I use most frequently, and in any case, I can just use the keyboard keys for this.