Better display Apple Music queries.
Display while querying as Apple’s servers are slow. If no matches, say so. If 1 match, just open it.

Currently only works for Internet radio stations but should be able to expand to Hermes without much difficulty in future.

As in the previous commit, don’t use current_stream_URL to check for an Internet radio station as it may be missing; just use current_stream_title directly.